+++ wtforms/validators.py
@@ -6,6 +6,10 @@ import uuid
 import warnings
 
 try:
+    import email_validator
+except ImportError:
+    email_validator = None
+try:
     import ipaddress
 except ImportError:
     ipaddress = None
@@ -293,12 +297,22 @@ class Regexp(object):
 
 class Email(object):
     """
-    Validates an email address. Note that this uses a very primitive regular
-    expression and should only be used in instances where you later verify by
-    other means, such as email activation or lookups.
+    Validates an email address. Requires email_validator package to be
+    installed. For ex: pip install wtforms[email].
 
     :param message:
         Error message to raise in case of a validation error.
+    :param granular_messsage:
+        Use validation failed message from email_validator library
+        (Default False).
+    :param check_deliverability:
+        Perform domain name resolution check (Default False).
+    :param allow_smtputf8:
+        Fail validation for addresses that would require SMTPUTF8
+        (Default True).
+    :param allow_empty_local:
+        Allow an empty local part (i.e. @example.com), e.g. for validating
+        Postfix aliases (Default False).
     """
 
     user_regex = re.compile(
@@ -306,28 +320,39 @@ class Email(object):
         r'|^"([\001-\010\013\014\016-\037!#-\[\]-\177]|\\[\001-\011\013\014\016-\177])*"\Z)',  # quoted-string
         re.IGNORECASE)
 
-    def __init__(self, message=None):
+    def __init__(
+        self,
+        message=None,
+        granular_message=False,
+        check_deliverability=False,
+        allow_smtputf8=True,
+        allow_empty_local=False,
+    ):
+        if email_validator is None:
+            raise Exception("Install 'email_validator' for email validation support.")
         self.message = message
-        self.validate_hostname = HostnameValidation(
-            require_tld=True,
-        )
+        self.granular_message = granular_message
+        self.check_deliverability = check_deliverability
+        self.allow_smtputf8 = allow_smtputf8
+        self.allow_empty_local = allow_empty_local
 
     def __call__(self, form, field):
-        value = field.data
-
-        message = self.message
-        if message is None:
-            message = field.gettext('Invalid email address.')
-
-        if not value or '@' not in value:
-            raise ValidationError(message)
-
-        user_part, domain_part = value.rsplit('@', 1)
-
-        if not self.user_regex.match(user_part):
-            raise ValidationError(message)
-
-        if not self.validate_hostname(domain_part):
+        try:
+            if field.data is None:
+                raise email_validator.EmailNotValidError()
+            email_validator.validate_email(
+                field.data,
+                check_deliverability=self.check_deliverability,
+                allow_smtputf8=self.allow_smtputf8,
+                allow_empty_local=self.allow_empty_local,
+            )
+        except email_validator.EmailNotValidError as e:
+            message = self.message
+            if message is None:
+                if self.granular_message:
+                    message = field.gettext(e)
+                else:
+                    message = field.gettext("Invalid email address.")
             raise ValidationError(message)